api/requirements.txt
openai>=1.0.0
python-dotenv
requests
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pandas
numpy
gspread
oauth2client
python-dateutil
faiss-cpu 